Hide the mail orer modal link button when there are no options, fetch pharmacy data outside for now until we iron this out
